Refactor, sync DB, and add script to sync test DB with dev DB
