+++ backend/package.json
@@ -4,6 +4,7 @@
   "private": true,
   "scripts": {
     "dev": "ts-node src/server.ts",
+    "init:test-db": "ts-node tests/initTestDB.ts",
     "seed": "ts-node prisma/seed.ts",
     "test": "vitest"
   },
