+++ backend/src/app.ts
@@ -1,7 +1,7 @@
 import { Hono } from 'hono';
 import { logger } from 'hono/logger';
 // import { cors } from 'hono/cors';
-import { news } from './routes/news';
+import { newsRoutes } from './controllers/news/routes';
 
 export const app = new Hono();
 
@@ -12,4 +12,4 @@ app.get('/', (c) => {
   return c.text('Wooo Health Check!')
 });
 
-app.route('/api/news', news);
+app.route('/api/news', newsRoutes);
