+++ backend/src/db/newsEmailRepository.ts
@@ -1,5 +1,5 @@
 import { prisma } from '../../prisma/client';
-import type { NewsEmailInput } from '../schemas/newsEmailSchema';
+import type { NewsEmailInput } from '../schemas/newsEmail.schema';
 
 export async function saveNewsEmail(data: NewsEmailInput) {
   return prisma.newsEmail.create({ data });
