+++ backend/src/schemas/newsEmail.schema.ts
