--- backend/tests/test.setup.ts
@@ -1,6 +1,0 @@
-import { config } from 'dotenv';
-import path from 'path';
-
-config({ path: path.resolve(__dirname, '../.env.test') });
-
-console.log(`Correctly testing against ${process.env.DATABASE_URL}.`);
