+++ backend/vitest.config.ts
@@ -2,8 +2,7 @@ import { defineConfig } from 'vitest/config';
 
 export default defineConfig({
   test: {
-    setupFiles: ['./tests/test.setup.ts'],
     globals: true,
-    environment: 'node',
+    environment: 'node'
   },
 });
